phuc forgot pass + change pass + sua lai login

+++ src/main/java/controllers/repository/JobSeekerRepository.java
@@ -3,9 +3,9 @@ package controllers.repository;
 import model.entity.JobSeeker;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
+    
 @Repository
-public interface JobSeekerRepository extends JpaRepository<JobSeeker, Long> {
+public interface JobSeekerRepository extends JpaRepository<JobSeeker, Integer> {
     JobSeeker findByEmail(String email);
 }
